+++ frontend/src/components/UI/Loader/Loader.tsx
@@ -1,0 +1,22 @@
+import styles from "./Loader.module.css";
+
+const Loader = () => {
+  return (
+    <div className={styles.loader}>
+      <div></div>
+      <div></div>
+      <div></div>
+      <div></div>
+      <div></div>
+      <div></div>
+      <div></div>
+      <div></div>
+      <div></div>
+      <div></div>
+      <div></div>
+      <div></div>
+    </div>
+  );
+};
+
+export default Loader;
