+++ frontend/src/components/forms/SubmitButton/SubmitButton.module.css
@@ -1,0 +1,9 @@
+.submitBtn {
+  min-height: 3.5rem;
+  margin-top: 1.8rem;
+
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  gap: 1rem;
+}
